ci: Run workflows only when relevant and add concurrency
Run the workflows only when they really need to run. E.g. markdownlint should only run when *.md
files are changed or the workflow file for markdownlint has been changed.

Additionally, adding concurrency to avoid multiple runs of the same workflow on the same commit.
This does not apply to all jobs, e.g. release.yml is excempt from that.

Signed-off-by: Steffen Scheib <steffen@scheib.me>